+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/building/repository/BuildingCustomRepositoryImpl.java
@@ -1,19 +1,21 @@
 package com.softeer5.uniro_backend.building.repository;
 
-import static com.softeer5.uniro_backend.building.entity.QBuilding.*;
-import static com.softeer5.uniro_backend.map.entity.QNode.node;
 
 import java.util.List;
 
+import com.softeer5.uniro_backend.building.service.vo.QBuildingNode;
 import org.springframework.stereotype.Repository;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.softeer5.uniro_backend.building.service.vo.QBuildingNode;
+
 import com.softeer5.uniro_backend.common.CursorPage;
 import com.softeer5.uniro_backend.building.service.vo.BuildingNode;
 
 import lombok.RequiredArgsConstructor;
+
+import static com.softeer5.uniro_backend.building.entity.QBuilding.building;
+import static com.softeer5.uniro_backend.map.entity.QNode.node;
 
 @Repository
 @RequiredArgsConstructor
